Conflicts:
	src/backend/bridge/dml/executor/plan_executor.cpp

Former-commit-id: e8a71ded5996c0722edd904f33c1e2e5eb7cf7d1

+++ src/backend/expression/parameter_value_expression.h
@@ -57,7 +57,7 @@ public:
     }
 
   private:
-    int m_valueIdx;
+    size_t m_valueIdx;
 
     Value m_paramValue;
 };
